move target grids into data folder
