+++ moose-etl/gcm-target-grid/pr/transform.py
@@ -8,7 +8,7 @@ working_grid_pp_path = "pp-data/*.pp"
 working_grid_nc_path = "moose_grid.nc"
 
 target_grid_path = Path(
-    "../../../src/ml_downscaling_emulator/utils/target-grids/60km/global/pr/moose_grid.nc"
+    "../../../src/ml_downscaling_emulator/data/target-grids/60km/global/pr/moose_grid.nc"
 )
 
 # convert pp data to netcdf and open with xr
