+++ src/ml_downscaling_emulator/data/target-grids/2.2km-coarsened-2x/uk/moose_grid.nc
