+++ src/ml_downscaling_emulator/data/target-grids/2.2km-coarsened-8x/uk/moose_grid.nc
