+++ src/ml_downscaling_emulator/data/target-grids/60km/global/pr/moose_grid.nc
