+++ src/ml_downscaling_emulator/data/target-grids/60km/global/vorticity850/moose_grid.nc
